improve logging with date check for old files only if there is a file in memory

+++ js5/4/assets.js
@@ -11,7 +11,7 @@ const findAll = (req, res) => {
 
 const findFile = (req, res) => {
   const filename = req.params.filename
-  console.log(`assets: fetching file ${filename}`)
+  console.log(`${new Date().toString()} assets: fetching file ${filename}`)
   if (!files.hasOwnProperty(filename)) return res.statusStatus(404)
   fs.readFile(path + filename, "utf-8", (err, data) => {
     if (err) {
@@ -24,6 +24,7 @@ const findFile = (req, res) => {
 const createFile = (req, res) => {
   const filename = req.body.filename
   const content = req.body.content
+  console.log(`${new Date().toString()} assets: writing file ${filename}`)
   fs.writeFile(path + filename, content, () => {
     files[filename] = Date.now()
     res.sendStatus(201)
@@ -31,16 +32,18 @@ const createFile = (req, res) => {
 }
 
 const deleteOldFiles = () => {
-  const maxTime = Date.now() - FILE_LIFETIME
-  console.log(`assets: checking for old files`)
-  Object.keys(files).forEach((file) => {
-    if (files[file] < maxTime) {
-      fs.unlink(path + file, () => {
-        console.log(`assets: deleted file ${file}`)
-        delete files[file]
-      })
-    }
-  })
+  if (Object.keys(files).length > 0) {
+    const maxTime = Date.now() - FILE_LIFETIME
+    console.log(`${new Date().toString()} assets: checking for old files`)
+    Object.keys(files).forEach((file) => {
+      if (files[file] < maxTime) {
+        fs.unlink(path + file, () => {
+          console.log(`${new Date().toString()} assets: deleted file ${file}`)
+          delete files[file]
+        })
+      }
+    })
+  }
 }
 
 const initialize = () => {
